requirements.txt
# Core dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.1.0
httpx>=0.27.0

# MCP client
mcp>=0.9.0

# 搜索源依赖
ddgs>=6.0.0  # duckduckgo-search 已重命名为 ddgs
wikipedia-api>=0.6.0  # 比 wikipedia 库更稳定

# Web 服务器依赖
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type checking (optional)
mypy>=1.5.0
types-pyyaml>=6.0.0
